Update Pi Instructions for Trixie (#1042)
* Update Pi Instructions for Trixie

Includes the newer `.sources` file format, and for the moment a comment and possible deletion of the dkms-xtra section. Wasn't needed at least on the 500+, we shall see if it's needed on the 500/5/CM5.

* Removed deb-mm source as not needed.

Removed the deb-mulimedia repo setup instructions as they are no longer needed. fdkaac can be installed from the stock repos. Thanks to @Taclane for the assistance.

* Update INSTALL-PI.md

Updated RAM recommendations.

* Update INSTALL-PI.md

---------

Co-authored-by: Luke Berndt <lukekb@gmail.com>