Standardize trunk-recorder.service & Contextualizes Comments & Commands. (#1072)
* Updates service file for systemd.

Uses the script by @taclane. Expects that the `trunk-recorder` binary is in `/usr/local/bin/` and that the user `trunk` exists. `LimitCORE=infinity` is the same as `ulimit -c` for core dumps.

* Add comments for core dump troubleshooting

Added comments for troubleshooting core dumps and usage instructions.

* Create auto-restart script for server

Sorta reverts the old auto-restart.sh but renames it so people can get a dead simple script to start off from.